refactor: rename tests and fix references in tests
